backend/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["eyerusalem <eyerusad12@gmail.com>"]
readme = "README.md"
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
quart = "^0.19.5"
openai = "^1.28.0"
python-dotenv = "^1.0.1"
langchain = "^0.1.19"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "api:run"